Removed compilation warnings on Windows

+++ Source/cg_LinkStrategies.cpp
@@ -245,7 +245,7 @@ void CircularFixedAngle::computeParameters_implementation(Sources const & finalS
 
     auto const primarySourceFinalPosition{ primarySourceFinalState.getPos() };
     mPrimarySourceFinalAngle = Radians::angleOf(primarySourceFinalPosition);
-    mDeviationPerSource = Degrees{ 360.0f } / finalStates.size();
+    mDeviationPerSource = Degrees{ 360.0f } / static_cast<float>(finalStates.size());
     auto const primarySourceInitialAngle{ Radians::angleOf(primarySourceInitialState.position) };
     mRotation = mPrimarySourceFinalAngle - primarySourceInitialAngle;
 
@@ -292,7 +292,7 @@ void CircularFixedAngle::enforce_implementation(Sources & finalStates,
     auto constexpr notQuiteZero{ 0.0000001f };
     auto const ordering{ mOrdering[sourceIndex.get()] };
 
-    auto const finalAngle{ mPrimarySourceFinalAngle + mDeviationPerSource * ordering };
+    auto const finalAngle{ mPrimarySourceFinalAngle + mDeviationPerSource * static_cast<float>(ordering) };
     auto const primaryDistFromOrig{ initialStates.primary.position.getDistanceFromOrigin() == 0.0f
                                         ? notQuiteZero
                                         : initialStates.primary.position.getDistanceFromOrigin() };
@@ -361,7 +361,7 @@ void CircularFullyFixed::computeParameters_implementation(Sources const & finalS
 
     auto const primarySourceFinalPosition{ primarySourceFinalState.getPos() };
     mPrimarySourceFinalAngle = Radians::angleOf(primarySourceFinalPosition);
-    mDeviationPerSource = Degrees{ 360.0f } / finalStates.size();
+    mDeviationPerSource = Degrees{ 360.0f } / static_cast<float>(finalStates.size());
     auto const primarySourceInitialAngle{ Radians::angleOf(primarySourceInitialState.position) };
     mRotation = mPrimarySourceFinalAngle - primarySourceInitialAngle;
 
@@ -406,7 +406,7 @@ void CircularFullyFixed::enforce_implementation(Sources & finalStates,
 {
     auto const ordering{ mOrdering[sourceIndex.get()] };
 
-    auto const finalAngle{ mPrimarySourceFinalAngle + mDeviationPerSource * ordering };
+    auto const finalAngle{ mPrimarySourceFinalAngle + mDeviationPerSource * static_cast<float>(ordering) };
     juce::Point<float> const finalPosition{ std::cos(finalAngle.getAsRadians()) * mRadius,
                                             std::sin(finalAngle.getAsRadians()) * mRadius };
 
@@ -556,7 +556,7 @@ SourceSnapshot FixedElevation::computeInitialStateFromFinalState_implementation(
 void LinearMin::computeParameters_implementation(Sources const & sources, SourcesSnapshots const & /*snapshots*/)
 {
     mBaseElevation = sources.getPrimarySource().getElevation();
-    mElevationPerSource = ELEVATION_DIFF / (sources.size() - 1);
+    mElevationPerSource = ELEVATION_DIFF / static_cast<float>((sources.size() - 1));
 }
 
 //==============================================================================
@@ -564,7 +564,7 @@ void LinearMin::enforce_implementation(Sources & finalStates,
                                        SourcesSnapshots const & /*initialStates*/,
                                        SourceIndex const sourceIndex) const
 {
-    auto const newElevation{ mBaseElevation + mElevationPerSource * sourceIndex.get() };
+    auto const newElevation{ mBaseElevation + mElevationPerSource * static_cast<float>(sourceIndex.get()) };
     finalStates[sourceIndex].setElevation(newElevation, Source::OriginOfChange::link);
 }
 
@@ -574,7 +574,7 @@ SourceSnapshot LinearMin::computeInitialStateFromFinalState_implementation([[may
                                                                            SourceIndex const sourceIndex) const
 {
     SourceSnapshot result{};
-    result.z = mBaseElevation - mElevationPerSource * sourceIndex.get();
+    result.z = mBaseElevation - mElevationPerSource * static_cast<float>(sourceIndex.get());
     return result;
 }
 
@@ -582,7 +582,7 @@ SourceSnapshot LinearMin::computeInitialStateFromFinalState_implementation([[may
 void LinearMax::computeParameters_implementation(Sources const & sources, SourcesSnapshots const & /*snapshots*/)
 {
     mBaseElevation = sources.getPrimarySource().getElevation();
-    mElevationPerSource = ELEVATION_DIFF / (sources.size() - 1);
+    mElevationPerSource = ELEVATION_DIFF / static_cast<float>((sources.size() - 1));
 }
 
 //==============================================================================
@@ -590,7 +590,7 @@ void LinearMax::enforce_implementation(Sources & finalStates,
                                        SourcesSnapshots const & /*initialStates*/,
                                        SourceIndex const sourceIndex) const
 {
-    auto const newElevation{ mBaseElevation + mElevationPerSource * sourceIndex.get() };
+    auto const newElevation{ mBaseElevation + mElevationPerSource * static_cast<float>(sourceIndex.get()) };
     finalStates[sourceIndex].setElevation(newElevation, Source::OriginOfChange::link);
 }
